Improve self voice plugin
Add a persistent tag to append to the spoken text to be able to have a persistent text in braille or to have a standard message displayed.
Known issue: script.presentMessage does not look like to provide a parameter to force voice interuption, so it is not possible to use the append tag with non-persistent message.